fastapi/requirements.txt
fastapi
uvicorn
oracledb
python-dotenv
requests
